Bug fix, v1.1
There has been an incredibly rare bug where the tool would just crush. The thing that caused it was the mistake in the behaviour of the compressor in case the new UncompressedData was longer than the original one.  
The bug is now fixed while truncating the 0s at the end of the last block.